deepvalue-app/backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-dotenv==0.19.0
sqlalchemy==1.4.23
pydantic==1.8.2
requests==2.26.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
aiohttp==3.8.1
pandas==1.3.3
numpy==1.21.2
yfinance==0.1.63
transformers==4.11.3
torch==1.9.1
